requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
gspread==6.1.4
google-auth==2.38.0
pydantic==2.10.6
python-dotenv==1.0.1